import from 2.70-alpha1 from http://downloads.sourceforge.net/project/mygifts/latest%20version/2.70-alpha1/myGifts-2.70-alpha1.tar.gz

+++ includes/Gift.class.php
@@ -7,6 +7,7 @@ class Gift {
 	var $forUser;
 	var $owner;
 	var $name;
+  var $category = 1;
 	var $comment;
 	var $url;
 	var $restricted;
@@ -14,26 +15,39 @@ class Gift {
 	var $price;
 	var $qty;
 	var $emptyObj;
-  
+
+  var $ts_creation;
+  var $ts_update;
+  var $new = false;
+  var $recent = false;
+
 	function Gift($id = "")
 	{
 		global $database, $user;
-		
+
 		$this->owner = $user->id;
 		if (!empty($id)) {
 			$this->id = (int) $id;
-			$database->loadObject($this, "select * from gft_gift where owner='".addslashes($this->owner)."' and id=".$this->id);
+			$database->loadObject($this, "select *, ".$database->timestamp("created")." as ts_created, ".$database->timestamp("updated")." as ts_updated from gft_gift where owner='".addslashes($this->owner)."' and id=".$this->id);
+      $curTime = time();
+      if ($curTime-$this->ts_created < 4*24*3600)
+        $this->new = true;
+      else if ($curTime-$this->ts_updated < 4*24*3600)
+        $this->recent = true;
 		} else
       $this->emptyObj = true;
+    if ($this->category === null)
+      $this->category = 0;
 	}
-	
-	function saveGift($forUser, $name, $comment, $url, $restricted, $offered, $offeredOn, $picture, $price, $priority, $qty)
+
+	function saveGift($forUser, $name, $category, $comment, $url, $restricted, $offered, $offeredOn, $picture, $price, $priority, $qty)
 	{
 		global $database;
 
 		$restricted = Tools::checkboxToInt($restricted);
 		$offered = Tools::checkboxToInt($offered);
 		$name = trim(stripslashes(strip_tags($name)));
+    $category = (int) $category;
 		$comment = trim(stripslashes(strip_tags($comment)));
 		$url = strtr(trim(strip_tags($url)), "\"\' ", "+++");
 		$picture = strtr(trim(strip_tags($picture)), "\"\' ", "+++");
@@ -43,23 +57,23 @@ class Gift {
 
 		if (empty($forUser))
 			$forUser = $this->owner;
-			
+
 		if (!empty($this->id)) {
-			$database->query("update gft_gift set name='".addslashes($name)."', comment='".addslashes($comment)."', url='".addslashes($url)."', restricted=$restricted, offered=$offered, offeredOn='".addslashes($offeredOn)."', picture='".addslashes($picture)."', price=$price, priority=$priority where id=".$this->id." and owner='".addslashes($this->owner)."'"); 
+			$database->query("update gft_gift set name='".addslashes($name)."', category=$category, comment='".addslashes($comment)."', url='".addslashes($url)."', restricted=$restricted, offered=$offered, offeredOn='".addslashes($offeredOn)."', picture='".addslashes($picture)."', price=$price, priority=$priority, updated=now() where id=".$this->id." and owner='".addslashes($this->owner)."'");
 		} else {
       $database->lock("gft_gift");
       $id = $database->getNextId("gft_gift");
-			$database->query("insert into gft_gift (id, forUser, owner, name, comment, url, restricted, offered, offeredOn, picture, price, priority) values($id, '".addslashes($forUser)."', '".addslashes($this->owner)."', '".addslashes($name)."', '".addslashes($comment)."', '".addslashes($url)."', $restricted, $offered, '".addslashes($offeredOn)."', '".addslashes($picture)."', $price, $priority)"); 
+			$database->query("insert into gft_gift (id, forUser, owner, name, category, comment, url, restricted, offered, offeredOn, picture, price, priority, created) values($id, '".addslashes($forUser)."', '".addslashes($this->owner)."', '".addslashes($name)."', $category, '".addslashes($comment)."', '".addslashes($url)."', $restricted, $offered, '".addslashes($offeredOn)."', '".addslashes($picture)."', $price, $priority, now())");
       $database->unlock("gft_gift");
     }
 		return 0;
 	}
-	
+
 	function removeGift()
 	{
 		global $database;
-		
-		$database->query("delete from gft_gift where id=".$this->id." and owner='".addslashes($this->owner)."'"); 
+
+		$database->query("delete from gft_gift where id=".$this->id." and owner='".addslashes($this->owner)."'");
 
 		return 0;
 	}
@@ -67,8 +81,11 @@ class Gift {
 	function claimGift($userId)
 	{
 		global $database;
-		
-		$database->query("insert into gft_claim (giftId, userId) values(".$this->id.", '".addslashes($userId)."')"); 
+
+    if (empty($userId))
+      return;
+
+		$database->query("insert into gft_claim (giftId, userId) values(".$this->id.", '".addslashes($userId)."')");
 
 		return 0;
 	}
@@ -76,21 +93,24 @@ class Gift {
 	function unclaimGift($userId)
 	{
 		global $database;
-		
-		$database->query("delete from gft_claim where giftId=".$this->id." and userId='".addslashes($userId)."'"); 
+
+    if (empty($userId))
+      return;
+
+		$database->query("delete from gft_claim where giftId=".$this->id." and userId='".addslashes($userId)."'");
 
 		return 0;
 	}
-	
+
 	function offeredGift($offeredOn)
 	{
 		global $database;
-		
-		$database->query("delete from gft_claim where giftId=".$this->id." and userId='".addslashes($userId)."'"); 
+
+		$database->query("delete from gft_claim where giftId=".$this->id." and userId='".addslashes($userId)."'");
 
 		return 0;
 	}
-	
+
 	function getParams()
 	{
 	}
@@ -98,7 +118,7 @@ class Gift {
 
 Controler::registerHandler("editGift", "display", "Gift", null, 1);
 Controler::registerHandler("modifyGift", "action", "Gift", array("id"), 1);
-Controler::registerHandler("saveGift", "action", "Gift", array("id", "forUser", "name", "comment", "url", "restricted", "offered", "offeredOn", "picture", "price", "priority", "qty"), 1);
+Controler::registerHandler("saveGift", "action", "Gift", array("id", "forUser", "name", "category", "comment", "url", "restricted", "offered", "offeredOn", "picture", "price", "priority", "qty"), 1);
 Controler::registerNextAction("saveGift", "close_refresh");
 Controler::registerHandler("removeGift", "action", "Gift", array("id"), 1);
 Controler::registerHandler("claimGift", "action", "Gift", array("id", "user"), 1);
